+++ src/PresentationalComponents/Breadcrumbs/_breadcrumbs.scss
@@ -1,0 +1,5 @@
+nav.pf-c-breadcrumb {
+    background-color: #fff;
+    padding: 10px;
+    li { text-transform: capitalize; }
+}
